+++ Changelog.md
@@ -3,6 +3,7 @@
 Language Features:
  * Allow accessing external functions via contract and interface names to obtain their selector.
  * Inline Assembly: Support literals ``true`` and ``false``.
+ * Allow interfaces to inherit from other interfaces
 
 
 Compiler Features: